move GenericMenu, GenericMenuItem and useHandleMennuAction into a folder
