+++ apps/meteor/client/components/GenericMenu/hooks/useHandleMenuAction.tsx
@@ -1,6 +1,6 @@
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 
-import type { GenericMenuItemProps } from '../components/GenericMenuItem';
+import type { GenericMenuItemProps } from '../GenericMenuItem';
 
 export const useHandleMenuAction = (items: GenericMenuItemProps[], close?: () => void) => {
 	return useMutableCallback((id) => {
