+++ apps/meteor/client/sidebar/header/actions/Administration.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { HTMLAttributes, VFC } from 'react';
 import React from 'react';
 
-import GenericMenu from '../../../components/GenericMenu';
+import GenericMenu from '../../../components/GenericMenu/GenericMenu';
 import { useAdministrationMenu } from './hooks/useAdministrationMenu';
 
 const Administration: VFC<Omit<HTMLAttributes<HTMLElement>, 'is'>> = (props) => {
