+++ apps/meteor/client/sidebar/header/actions/CreateRoom.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { HTMLAttributes, VFC } from 'react';
 import React from 'react';
 
-import GenericMenu from '../../../components/GenericMenu';
+import GenericMenu from '../../../components/GenericMenu/GenericMenu';
 import { useCreateRoom } from './hooks/useCreateRoomMenu';
 
 const CreateRoom: VFC<Omit<HTMLAttributes<HTMLElement>, 'is'>> = (props) => {
