+++ apps/meteor/client/sidebar/header/actions/Sort.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { VFC, HTMLAttributes } from 'react';
 import React from 'react';
 
-import GenericMenu from '../../../components/GenericMenu';
+import GenericMenu from '../../../components/GenericMenu/GenericMenu';
 import { useSortMenu } from './hooks/useSortMenu';
 
 const Sort: VFC<Omit<HTMLAttributes<HTMLElement>, 'is'>> = (props) => {
